# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "رسالة الخيرية API",
    "description": "## نظام إدارة رسالة الخيرية\n\n### الأدوار المتاحة:\n| الدور | طريقة الدخول | الصلاحيات |\n|-------|-------------|----------|\n| **Donor** | Phone Number | تسجيل — عرض بياناته — الاشتراك |\n| **Reception** | Username | إدارة المتبرعين — تسجيل تبرعات |\n| **Admin** | Username | كل الصلاحيات + إدارة الموظفين |\n\n### كيفية الاستخدام:\n1. سجّل دخول من `/api/v1/auth/login`\n2. انسخ الـ `token` من الـ Response\n3. اضغط **Authorize** في الأعلى والصق: `Bearer {token}`",
    "contact": {
      "name": "Resala Dev Team",
      "email": "dev@resala.org"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/emergency-cases": {
      "post": {
        "tags": [
          "EmergencyCases — Admin"
        ],
        "summary": "[Admin] إنشاء حالة حرجة جديدة",
        "description": "يُنشئ حالة حرجة جديدة في النظام.\r\n            \r\n**⚠️ يتطلب Bearer Token بدور Admin**\r\n            \r\n**Request Example:**\r\n```json\r\n{\r\n  \"title\": \"عملية عاجلة\",\r\n  \"description\": \"مريض يحتاج عملية فورًا\",\r\n  \"targetAmount\": 50000\r\n}\r\n```\r\n            \r\n**Success Response (201):**\r\n```json\r\n{\r\n  \"id\": 1,\r\n  \"title\": \"عملية عاجلة\",\r\n  \"description\": \"مريض يحتاج عملية فورًا\",\r\n  \"targetAmount\": 50000,\r\n  \"collectedAmount\": 0,\r\n  \"createdOn\": \"2026-03-25T00:00:00\"\r\n}\r\n```",
        "operationId": "EmergencyCases_Create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "UrgencyLevel": {
                    "$ref": "#/components/schemas/UrgencyLevel"
                  },
                  "RequiredAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "Attachment": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "UrgencyLevel": {
                  "style": "form"
                },
                "RequiredAmount": {
                  "style": "form"
                },
                "Attachment": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyCaseViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EmergencyCases — Public"
        ],
        "summary": "جلب كل الحالات الحرجة",
        "description": "يرجع قائمة بكل الحالات الحرجة المسجلة في النظام.",
        "operationId": "EmergencyCases_GetAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmergencyCaseViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emergency-cases/{id}": {
      "get": {
        "tags": [
          "EmergencyCases — Public"
        ],
        "summary": "جلب حالة حرجة بالـ ID",
        "description": "عرض تفاصيل حالة حرجة باستخدام المعرف",
        "operationId": "EmergencyCases_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم الحالة الحرجة",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تفاصيل الحالة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyCaseViewModel"
                }
              }
            }
          },
          "404": {
            "description": "الحالة غير موجودة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EmergencyCases — Admin"
        ],
        "summary": "[Admin] تعديل حالة حرجة",
        "description": "تعديل بيانات حالة حرجة موجودة.\r\n            \r\n**⚠️ يتطلب Bearer Token بدور Admin**",
        "operationId": "EmergencyCases_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم الحالة الحرجة",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "UrgencyLevel": {
                    "$ref": "#/components/schemas/UrgencyLevel"
                  },
                  "RequiredAmount": {
                    "type": "number",
                    "format": "double"
                  },
                  "Attachment": {
                    "type": "string",
                    "format": "binary"
                  },
                  "IsActive": {
                    "type": "boolean"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "UrgencyLevel": {
                  "style": "form"
                },
                "RequiredAmount": {
                  "style": "form"
                },
                "Attachment": {
                  "style": "form"
                },
                "IsActive": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyCaseViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EmergencyCases — Admin"
        ],
        "summary": "[Admin] حذف حالة حرجة",
        "description": "حذف حالة حرجة من النظام — يتطلب Admin",
        "operationId": "EmergencyCases_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "رقم الحالة الحرجة",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "تم الحذف بنجاح"
          },
          "404": {
            "description": "الحالة غير موجودة",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmergencyCaseViewModel": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "nullable": true
          },
          "imagePublicId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "targetAmount": {
            "type": "number",
            "format": "double"
          },
          "receivedAmount": {
            "type": "number",
            "format": "double"
          },
          "urgencyLevel": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UrgencyLevel": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "أدخل الـ Token هكذا: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}